+++ tests/Event/CakeHttpFlowEventTest.php
@@ -3,34 +3,43 @@
 namespace PennyTest\Event;
 
 use Exception;
-use Penny\Event\HttpFlowEvent;
+use Penny\Event\CakeHttpFlowEvent;
 use Penny\Route\RouteInfoInterface;
 use PHPUnit_Framework_TestCase;
 use Zend\Diactoros\Request;
 use Zend\Diactoros\Response;
 use Zend\Diactoros\Uri;
 
-class HttpFlowEventTest extends PHPUnit_Framework_TestCase
+class CakeHttpFlowEventTest extends PHPUnit_Framework_TestCase
 {
-    private $event;
+    /** @var CakeEvent */
+    protected $event;
 
-    public function setUp()
+    protected function setUp()
     {
-        $request = (new Request())
-        ->withUri(new Uri('/'))
-        ->withMethod('GET');
-        $response = new Response();
+        $this->event = new CakeHttpFlowEvent('foo');
+    }
 
-        $this->event = new HttpFlowEvent('http_flow_event', $request, $response);
+    public function testGetName()
+    {
+        $this->assertEquals('foo', $this->event->getName());
     }
 
     public function testGetResponse()
     {
+        $response = new Response();
+        $this->event->setResponse($response);
+
         $this->assertInstanceOf(Response::class, $this->event->getResponse());
     }
 
     public function testGetRequest()
     {
+        $request = (new Request())
+        ->withUri(new Uri('/'))
+        ->withMethod('GET');
+        $this->event->setRequest($request);
+
         $this->assertInstanceOf(Request::class, $this->event->getRequest());
     }
 
@@ -47,5 +56,11 @@ class HttpFlowEventTest extends PHPUnit_Framework_TestCase
         $exception = new Exception();
         $this->event->setException($exception);
         $this->assertSame($exception, $this->event->getException());
+    }
+
+    public function testStopPropagation()
+    {
+        $this->event->stopPropagation(false);
+        $this->assertFalse($this->event->isStopped());
     }
 }
